requirements.txt
beautifulsoup4==4.12.2
browsermob-proxy==0.8.0
chromedriver-autoinstaller==0.4.0
fqdn==1.5.1
gspread==5.10.0
ipykernel==6.27.1
jsonpointer==2.3
mitmproxy==8.0.0
openpyxl==3.1.2
pandas==2.0.2
pip-chill==1.0.3
psycopg2-binary==2.9.7
pyarrow==12.0.1
pygsheets==2.0.6
pysocks==1.7.1
selenium==4.10.0
sqlalchemy==2.0.16
tqdm==4.65.0
uri-template==1.2.0
webcolors==1.13
webdriver-manager==4.0.0
wget==3.2
